requirements.txt
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
entrypoints==0.3
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.24
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
matplotlib==3.4.2
opencv-python==4.6.0.66
pandas==1.2.4
Pillow==8.2.0
pipreqs==0.4.10
py==1.10.0
pytest-shutil==1.7.0
python-utils==2.5.6
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
torch==1.7.1
torchvision==0.1.8
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6